# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DataProviderService",
        "version": "0.1"
    },
    "paths": {
        "/api/v1/transactions": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Get all transactions with filters",
                "description": "Get all transactions with filters",
                "operationId": "getTransactions",
                "parameters": [
                    {
                        "name": "statusCode",
                        "in": "query",
                        "description": "status code",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "paid",
                                "pending",
                                "reject"
                            ]
                        }
                    },
                    {
                        "name": "currency",
                        "in": "query",
                        "description": "currency",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "amountMin",
                        "in": "query",
                        "description": "minimum amount",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "amountMax",
                        "in": "query",
                        "description": "maximum amount",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "provider",
                        "in": "query",
                        "description": "provider",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "DataProviderX",
                                "DataProviderY",
                                "DataProviderW"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "payload": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "amount": {
                                                        "type": "number",
                                                        "example": 200
                                                    },
                                                    "currency": {
                                                        "type": "string",
                                                        "example": "USD"
                                                    },
                                                    "statusCode": {
                                                        "type": "string",
                                                        "example": "paid"
                                                    },
                                                    "provider": {
                                                        "type": "string",
                                                        "example": "DataProviderX"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2021-01-01 00:00:00"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        }
    }
}